+++ src/test/converter2/issues/gh2700.ts
@@ -1,0 +1,6 @@
+/**
+ * {@link Map.size | size user specified}
+ * {@link Map.size user specified}
+ * {@link Map.size}
+ */
+export const abc = new Map<string, number>();